+++ src/endstone/runtime/bedrock_hooks/script_player_gameplay_handler.cpp
@@ -33,7 +33,6 @@
 #include "endstone/core/player.h"
 #include "endstone/core/server.h"
 #include "endstone/event/actor/player_death_event.h"
-#include "endstone/event/player/player_arm_swing_event.h"
 #include "endstone/event/player/player_dimension_change_event.h"
 #include "endstone/event/player/player_game_mode_change_event.h"
 #include "endstone/event/player/player_interact_actor_event.h"
@@ -146,20 +145,6 @@ bool handleEvent(const PlayerDimensionChangeAfterEvent &event)
     return true;
 }
 
-bool handleEvent(const PlayerSwingStartEvent &event)
-{
-    if (const auto *player = event.player.tryUnwrap<::Player>(); player) {
-        const auto &server = endstone::core::EndstoneServer::getInstance();
-        std::optional<endstone::ItemStack> item;
-        if (!event.held_item.isNull()) {
-            item = endstone::core::EndstoneItemStack::fromMinecraft(event.held_item);
-        }
-        endstone::PlayerArmSwingEvent e{player->getEndstoneActor<endstone::core::EndstonePlayer>(), std::move(item)};
-        server.getPluginManager().callEvent(e);
-    }
-    return true;
-}
-
 bool handleEvent(const PlayerGetExperienceOrbEvent &event)
 {
     const auto &server = endstone::core::EndstoneServer::getInstance();
@@ -254,8 +239,7 @@ HandlerResult ScriptPlayerGameplayHandler::handleEvent1(const PlayerGameplayEven
                       std::is_same_v<T, Details::ValueOrRef<const PlayerFormResponseEvent>> ||
                       std::is_same_v<T, Details::ValueOrRef<const PlayerFormCloseEvent>> ||
                       std::is_same_v<T, Details::ValueOrRef<const ::PlayerRespawnEvent>> ||
-                      std::is_same_v<T, Details::ValueOrRef<const PlayerDimensionChangeAfterEvent>> ||
-                      std::is_same_v<T, Details::ValueOrRef<const PlayerSwingStartEvent>>) {
+                      std::is_same_v<T, Details::ValueOrRef<const PlayerDimensionChangeAfterEvent>>) {
             if (!handleEvent(arg.value())) {
                 return HandlerResult::BypassListeners;
             }